+++ src/modules/tools/stateTools.js
@@ -9,6 +9,7 @@ import SaveSelection from "./saveSelection/components/SaveSelection.vue";
 import ScaleSwitcher from "./scaleSwitcher/components/ScaleSwitcher.vue";
 import StyleVT from "./styleVT/components/StyleVT.vue";
 import SupplyCoord from "./supplyCoord/components/SupplyCoord.vue";
+import SelectFeatures from "./selectFeatures/components/SelectFeatures.vue";
 import BufferAnalysis from "./bufferAnalysis/components/BufferAnalysis.vue";
 
 /**
@@ -30,6 +31,7 @@ const state = {
         scaleSwitcher: ScaleSwitcher,
         styleVT: StyleVT,
         supplyCoord: SupplyCoord,
+        selectFeatures: SelectFeatures,
         bufferAnalysis: BufferAnalysis
     },
     configuredTools: []